qanda파일 생성, 앱 옮기기
