+++ csecom/common/forms.py
@@ -1,7 +1,6 @@
 from django import forms
 from django.contrib.auth.forms import ReadOnlyPasswordHashField
 from .models import User
-from .models import Qanda #Qanda models가져오기
 
 class UserCreationForm(forms.ModelForm):
     student_id = forms.CharField(label='학번', max_length=10)
@@ -68,8 +67,3 @@ class LoginForm(forms.Form):
         except User.DoesNotExist:   
             self.add_error("student_id", forms.ValidationError("이 등록되어있지 않습니다."))
 
-#qanda의 form부분
-class QandaForm(forms.ModelForm):
-    class Meta:
-        model = Qanda
-        fields = ['name', 'title', 'contents', 'url', 'email']
