+++ csecom/common/models.py
@@ -56,11 +56,3 @@ class User(AbstractBaseUser):
     @property
     def is_check(self):
         return self.is_student
-
-class Qanda(models.Model):
-    name = models.CharField(max_length=50)
-    title = models.CharField(max_length=50)
-    contents = models.TextField()
-    url = models.URLField()
-    email = models.EmailField()
-    cdate = models.DateTimeField(auto_now_add=True)
