+++ csecom/common/urls.py
@@ -8,6 +8,4 @@ urlpatterns = [
     path('login/', views.LoginView.as_view(), name='login'),
     path('logout/', views.log_out, name='logout'),
     path('signup/', views.signup, name='signup'),
-    path('qanda/', views.QandaView, name='qanda'),
-    path('qandalist/', views.QandalistView, name='qandalist'),
 ]
