+++ csecom/csecom/urls.py
@@ -12,5 +12,5 @@ urlpatterns = [
     path('locker/', include('locker.urls')),
     path('common/', include('common.urls')),
     path('si_checker/', include('si_checker.urls')),
-    path('qanda/', views.temp_qanda, name='qanda'), # 임시
+    path('qanda/', include('qanda.urls')), 
 ]
