+++ csecom/csecom/views.py
@@ -8,6 +8,3 @@ def empty_page(request):
 
 def test_page(request):
     return render(request, 'test.html')
-
-def temp_qanda(request):
-    return render(request, 'qanda.html')
