+++ csecom/qanda/forms.py
@@ -1,0 +1,7 @@
+from django import forms
+from .models import *
+
+class QandaForm(forms.ModelForm):
+    class Meta:
+        model = Qanda
+        fields = ['name', 'title', 'contents', 'url', 'email']
