+++ csecom/qanda/models.py
@@ -1,0 +1,10 @@
+from django.db import models
+
+class Qanda(models.Model):
+    name = models.CharField(max_length=50)
+    title = models.CharField(max_length=50)
+    contents = models.TextField()
+    url = models.URLField()
+    email = models.EmailField()
+    cdate = models.DateTimeField(auto_now_add=True)
+
